third_party/WebKit/LayoutTests/fast/css/local-file-name-requirements.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel="stylesheet" href="./resources/green.css"></link>
<link rel="stylesheet" href="./resources/red.not-css"></link>
<body>
  <div></div>
  <script>
    test(_ => {
      assert_equals(getComputedStyle(document.querySelector('div')).backgroundColor, "rgb(0, 128, 0)", 'green.css should be applied');
    }, "`red.not-css` should not load");
  </script>
</body>
